# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/complaints": {
      "post": {
        "operationId": "ComplaintsController_create",
        "parameters": [
          {
            "name": "X-Client-Transport",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "cookie"
              ]
            },
            "description": "Web panels send `cookie` on /auth/otp/verify, /auth/merchant/login, /auth/admin/login, and /auth/refresh to receive the refresh token ONLY as an httpOnly cookie, stripped out of the JSON body. Omit entirely for the mobile app (reads the refresh token from the body)."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateComplaintDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplaintTicketDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but not permitted (wrong actor type, unapproved merchant, or ownership check failed).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "File a complaint (ETAHS 15-calendar-day SLA starts now).",
        "tags": [
          "complaints"
        ]
      }
    },
    "/api/complaints/mine": {
      "get": {
        "operationId": "ComplaintsController_listMine",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "OPEN",
                "MERCHANT_RESPONDED",
                "RESOLVED",
                "ESCALATED"
              ],
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "X-Client-Transport",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "cookie"
              ]
            },
            "description": "Web panels send `cookie` on /auth/otp/verify, /auth/merchant/login, /auth/admin/login, and /auth/refresh to receive the refresh token ONLY as an httpOnly cookie, stripped out of the JSON body. Omit entirely for the mobile app (reads the refresh token from the body)."
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplaintListResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but not permitted (wrong actor type, unapproved merchant, or ownership check failed).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the caller's own complaints, paginated.",
        "tags": [
          "complaints"
        ]
      }
    },
    "/api/complaints/{id}": {
      "get": {
        "operationId": "ComplaintsController_getMine",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Client-Transport",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "cookie"
              ]
            },
            "description": "Web panels send `cookie` on /auth/otp/verify, /auth/merchant/login, /auth/admin/login, and /auth/refresh to receive the refresh token ONLY as an httpOnly cookie, stripped out of the JSON body. Omit entirely for the mobile app (reads the refresh token from the body)."
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplaintDetailResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but not permitted (wrong actor type, unapproved merchant, or ownership check failed).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get one of the caller's own complaints, with its message thread.",
        "tags": [
          "complaints"
        ]
      }
    },
    "/api/complaints/{id}/messages": {
      "post": {
        "operationId": "ComplaintsController_addMessage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Client-Transport",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "cookie"
              ]
            },
            "description": "Web panels send `cookie` on /auth/otp/verify, /auth/merchant/login, /auth/admin/login, and /auth/refresh to receive the refresh token ONLY as an httpOnly cookie, stripped out of the JSON body. Omit entirely for the mobile app (reads the refresh token from the body)."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateComplaintMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplaintMessageDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but not permitted (wrong actor type, unapproved merchant, or ownership check failed).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Post to a complaint's thread (CONSUMER owner, assigned MERCHANT, or ADMIN). A MERCHANT reply also flips OPEN -> MERCHANT_RESPONDED.",
        "tags": [
          "complaints"
        ]
      }
    },
...
    "/api/admin/complaints/{id}": {
      "get": {
        "operationId": "AdminComplaintsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Client-Transport",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "cookie"
              ]
            },
            "description": "Web panels send `cookie` on /auth/otp/verify, /auth/merchant/login, /auth/admin/login, and /auth/refresh to receive the refresh token ONLY as an httpOnly cookie, stripped out of the JSON body. Omit entirely for the mobile app (reads the refresh token from the body)."
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplaintDetailResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but not permitted (wrong actor type, unapproved merchant, or ownership check failed).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get one complaint (full detail + message thread).",
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Kurtar API",
    "description": "Kurtar — surplus-food marketplace API. The contract for all four\nfrontends (merchant-web, admin-web, the consumer app, and the\nlanding site).\n\n**Auth**: every route requires a bearer access token\n(`Authorization: Bearer <token>`) unless explicitly marked \"no auth\nrequired\" below — obtained via one of the /auth/* endpoints. A web\npanel additionally sends `X-Client-Transport: cookie` on every\nlogin/verify/refresh call to receive the refresh token ONLY as an\nhttpOnly cookie (never in the JSON body) — the mobile app omits\nthis header and reads the refresh token from the response body\ninstead.\n\n**Errors**: every error response is `{statusCode, errorCode,\nmessage}` (ErrorEnvelope, below) — branch on `errorCode`, never on\n`message` (not stable across releases, and may be Turkish or\nEnglish depending on the request).",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ErrorEnvelopeDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400,
            "description": "HTTP status code, duplicated from the response status for convenience."
          },
          "errorCode": {
            "type": "string",
            "example": "VALIDATION_ERROR",
            "description": "A stable, machine-readable identifier for this specific error condition — the field every client should branch on, never `message`."
          },
          "message": {
            "type": "string",
            "example": "overallStars must be between 1 and 5.",
            "description": "Human-readable detail, Turkish or English depending on the request — not stable across releases, do not pattern-match on it."
          }
        },
        "required": [
          "statusCode",
          "errorCode",
          "message"
        ]
      },
...
      "ComplaintTicketDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "merchantId": {
            "type": "string",
            "nullable": true
          },
          "reservationId": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "type": "string",
            "enum": [
              "FOOD_QUALITY",
              "MISSING_ITEMS",
              "WRONG_ITEMS",
              "STORE_CLOSED_NO_SHOW",
              "RUDE_STAFF",
              "PAYMENT_BILLING",
              "SAFETY_HYGIENE",
              "OTHER"
            ]
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "OPEN",
              "MERCHANT_RESPONDED",
              "RESOLVED",
              "ESCALATED"
            ]
          },
          "slaDeadlineAt": {
            "format": "date-time",
            "type": "string"
          },
          "resolvedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "slaWarningSentAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "userId",
          "category",
          "description",
          "status",
          "slaDeadlineAt",
          "createdAt",
          "updatedAt"
        ]
      },
      "ComplaintListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ComplaintTicketDto"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "pageSize"
        ]
      },
      "CreateComplaintDto": {
        "type": "object",
        "properties": {
          "category": {
            "type": "string",
            "enum": [
              "FOOD_QUALITY",
              "MISSING_ITEMS",
              "WRONG_ITEMS",
              "STORE_CLOSED_NO_SHOW",
              "RUDE_STAFF",
              "PAYMENT_BILLING",
              "SAFETY_HYGIENE",
              "OTHER"
            ]
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "maxLength": 4000
          },
          "reservationId": {
            "type": "string",
            "description": "If set, merchantId is derived from the reservation's store (any client-supplied merchantId is ignored)."
          },
          "merchantId": {
            "type": "string"
          }
        },
        "required": [
          "category",
          "description"
        ]
      },
      "ComplaintMessageDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "complaintId": {
            "type": "string"
          },
          "authorType": {
            "type": "string",
            "enum": [
              "CONSUMER",
              "MERCHANT",
              "ADMIN"
            ]
          },
          "authorId": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "complaintId",
          "authorType",
          "authorId",
          "body",
          "createdAt"
        ]
      },
      "ComplaintDetailResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "merchantId": {
            "type": "string",
            "nullable": true
          },
          "reservationId": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "type": "string",
            "enum": [
              "FOOD_QUALITY",
              "MISSING_ITEMS",
              "WRONG_ITEMS",
              "STORE_CLOSED_NO_SHOW",
              "RUDE_STAFF",
              "PAYMENT_BILLING",
              "SAFETY_HYGIENE",
              "OTHER"
            ]
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "OPEN",
              "MERCHANT_RESPONDED",
              "RESOLVED",
              "ESCALATED"
            ]
          },
          "slaDeadlineAt": {
            "format": "date-time",
            "type": "string"
          },
          "resolvedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "slaWarningSentAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ComplaintMessageDto"
            }
          }
        },
        "required": [
          "id",
          "userId",
          "category",
          "description",
          "status",
          "slaDeadlineAt",
          "createdAt",
          "updatedAt",
          "messages"
        ]
      },
      "CreateComplaintMessageDto": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string",
            "minLength": 1,
            "maxLength": 4000
          }
        },
        "required": [
          "body"
        ]
      },
...
    }
  }
}